frontend/requirements.txt
# Streamlit Frontend Requirements

# Core Framework
streamlit==1.29.0
streamlit-option-menu==0.3.6
streamlit-aggrid==0.3.4.post3
streamlit-extras==0.3.6

# API Communication
requests==2.31.0
httpx==0.25.2

# Data Handling
pandas==2.1.4
numpy==1.26.2

# Visualization
plotly==5.18.0
altair==5.2.0
matplotlib==3.8.2

# UI Components
pillow==10.1.0
streamlit-authenticator==0.2.3
streamlit-lottie==0.0.5
streamlit-toggle-switch==1.0.2

# Forms & Validation
pydantic==2.5.2
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
pytz==2023.3
